requirements.txt
aiohttp==3.7.4.post0
ansi==0.2.0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorlog==5.0.1
cryptography==3.4.7
daemonize==2.5.0
deepmerge==0.3.0
dulwich==0.20.23
errbot==6.1.8
Flask==2.0.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
multidict==5.1.0
pycparser==2.20
Pygments==2.9.0
pygments-markdown-lexer==0.1.0.dev39
pyOpenSSL==20.0.1
python-dotenv==0.17.1
requests==2.25.1
schedule==1.1.0
six==1.16.0
slackclient==1.3.2
soupsieve==2.2.1
typing-extensions==3.10.0.0
urllib3==1.26.5
waitress==2.0.0
WebOb==1.8.7
websocket-client==0.54.0
WebTest==2.0.35
Werkzeug==2.0.1
yarl==1.6.3
